# api/openapi.yaml
openapi: 3.0.3
info:
  title: Recommendation Service
  version: 0.1.0

paths:
  /catalogs/register:
    post:
      summary: Register a new catalog and obtain a secret key
      responses:
        '201':
          description: Catalog created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogWithKey'
        '500':
          $ref: '#/components/responses/InternalError'

  /catalogs/{catalogId}/ratings:
    put:
      summary: Add or modify a user rating
      security:
        - catalogKey: []
      parameters:
        - $ref: '#/components/parameters/catalogId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Rating'
      responses:
        '200':
          description: Rating saved
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Catalog or product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

  /catalogs/{catalogId}/ratings/{userId}/{productId}:
    delete:
      summary: Delete a user rating
      security:
        - catalogKey: []
      parameters:
        - $ref: '#/components/parameters/catalogId'
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Rating deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Catalog or product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

  /catalogs/{catalogId}/ratings/bulk:
    put:
      summary: Bulk add or modify ratings
      security:
        - catalogKey: []
      parameters:
        - $ref: '#/components/parameters/catalogId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Rating'
              minItems: 1
              maxItems: 2000
      responses:
        '200':
          description: Bulk operation completed — saved ratings and skipped failures
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkRatingResult'
        '400':
          description: Validation failed (e.g. exceeds 2000 items, invalid rating data)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/CatalogNotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

  /catalogs/{catalogId}/users/{userId}/recommendations:
    get:
      summary: Get recommendations for a user
      security:
        - catalogKey: []
      parameters:
        - $ref: '#/components/parameters/catalogId'
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of recommended products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recommendation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Catalog or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

  /catalogs/{catalogId}/products:
    post:
      summary: Create a new product
      security:
        - catalogKey: []
      parameters:
        - $ref: '#/components/parameters/catalogId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/CatalogNotFound'
        '409':
          description: A product with this ID already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

  /catalogs/{catalogId}/products/bulk:
    post:
      summary: Bulk-create products
      security:
        - catalogKey: []
      parameters:
        - $ref: '#/components/parameters/catalogId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Product'
              minItems: 1
              maxItems: 2000
      responses:
        '201':
          description: Bulk operation completed — created products and skipped duplicates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkProductResult'
        '400':
          description: Bad request (e.g. exceeds 2000 items, invalid batch size)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation failed (e.g. empty categories, invalid product data)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/CatalogNotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /catalogs/{catalogId}/products/{productId}/similar:
    get:
      summary: Get similar products for a given product
      security:
        - catalogKey: []
      parameters:
        - $ref: '#/components/parameters/catalogId'
        - $ref: '#/components/parameters/productId'
        - name: limit
          in: query
          required: false
          description: Maximum number of similar products to return
          schema:
            type: integer
            minimum: 1
            default: 10
        - name: strategy
          in: query
          required: false
          description: Similarity strategy to use
          schema:
            $ref: '#/components/schemas/SimilarityStrategy'
      responses:
        '200':
          description: Ranked list of similar products (empty list if no similarity data)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recommendation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Catalog or product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

  /catalogs/{catalogId}/products/{productId}:
    put:
      summary: Update a product's name or categories
      security:
        - catalogKey: []
      parameters:
        - $ref: '#/components/parameters/catalogId'
        - $ref: '#/components/parameters/productId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Catalog or product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      summary: Delete a product and gracefully remove its associated ratings
      security:
        - catalogKey: []
      parameters:
        - $ref: '#/components/parameters/catalogId'
        - $ref: '#/components/parameters/productId'
      responses:
        '204':
          description: Product deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Catalog or product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

components:
  responses:
    Unauthorized:
      description: Secret key missing or mismatched
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    CatalogNotFound:
      description: Catalog not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ProductNotFound:
      description: Product not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation failed (e.g. missing required header, invalid field values)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  securitySchemes:
    catalogKey:
      type: apiKey
      in: header
      name: X-Catalog-Key

  parameters:
    catalogId:
      name: catalogId
      in: path
      required: true
      schema:
        type: string
    productId:
      name: productId
      in: path
      required: true
      schema:
        type: string

  schemas:
    CatalogWithKey:
      type: object
      properties:
        catalogId:
          type: string
        secretKey:
          type: string

    Product:
      type: object
      required: [productId, name, categories]
      properties:
        productId:
          type: string
        name:
          type: string
        categories:
          type: array
          items:
            type: string
          minItems: 1

    BulkProductResult:
      type: object
      required: [created, skipped]
      properties:
        created:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        skipped:
          type: array
          items:
            $ref: '#/components/schemas/BulkSkippedProduct'

    BulkSkippedProduct:
      type: object
      required: [productId, reason]
      properties:
        productId:
          type: string
        reason:
          type: string

    BulkRatingResult:
      type: object
      required: [saved, skipped]
      properties:
        saved:
          type: array
          items:
            $ref: '#/components/schemas/Rating'
        skipped:
          type: array
          items:
            $ref: '#/components/schemas/BulkSkippedRating'

    BulkSkippedRating:
      type: object
      required: [userId, productId, reason]
      properties:
        userId:
          type: string
        productId:
          type: string
        reason:
          type: string

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        categories:
          type: array
          items:
            type: string
          minItems: 1

    Rating:
      type: object
      required: [userId, productId, score]
      properties:
        userId:
          type: string
        productId:
          type: string
        score:
          type: number
          format: float

    Recommendation:
      type: object
      properties:
        productId:
          type: string
        score:
          type: number
          format: float

    SimilarityStrategy:
      type: string
      enum:
        - auto
        - jaccard
        - euclidean
        - pearson
        - cosine
        - matrix_factorization
      default: auto

    Error:
      type: object
      required: [reason]
      properties:
        reason:
          type: string