# app/schemas.py
# generated by datamodel-codegen:
#   filename:  openapi.yaml
#   timestamp: 2026-04-20T01:06:49+00:00

from __future__ import annotations

from enum import Enum

from pydantic import BaseModel, Field


class CatalogWithKey(BaseModel):
    catalogId: str | None = None
    secretKey: str | None = None


class Product(BaseModel):
    productId: str
    name: str
    categories: list[str] = Field(..., min_length=1)


class BulkSkippedProduct(BaseModel):
    productId: str
    reason: str


class BulkSkippedRating(BaseModel):
    userId: str
    productId: str
    reason: str


class ProductUpdate(BaseModel):
    name: str | None = None
    categories: list[str] | None = Field(None, min_length=1)


class Rating(BaseModel):
    userId: str
    productId: str
    score: float


class Recommendation(BaseModel):
    productId: str | None = None
    score: float | None = None


class SimilarityStrategy(Enum):
    auto = 'auto'
    jaccard = 'jaccard'
    euclidean = 'euclidean'
    pearson = 'pearson'
    cosine = 'cosine'
    matrix_factorization = 'matrix_factorization'


class Error(BaseModel):
    reason: str


class BulkProductResult(BaseModel):
    created: list[Product]
    skipped: list[BulkSkippedProduct]


class BulkRatingResult(BaseModel):
    saved: list[Rating]
    skipped: list[BulkSkippedRating]